テストが通らなかった問題を修正
